feat: rename modules folders
Signals, Services and State namespaced by registerd module name.
Renaming would help to find it's representation on filesystem.
