+++ packages/cerebral-todomvc/src/modules/app/modules/footer/actions/clearCompleted.js
