+++ packages/cerebral-todomvc/src/modules/app/modules/footer/actions/setFilter.js
