+++ packages/cerebral-todomvc/src/modules/app/modules/footer/chains/clearCompleted.js
