+++ packages/cerebral-todomvc/src/modules/app/modules/footer/chains/setFilter.js
