+++ packages/cerebral-todomvc/src/modules/app/modules/footer/index.js
