+++ packages/cerebral-todomvc/src/modules/app/modules/list/actions/removeTodo.js
