+++ packages/cerebral-todomvc/src/modules/app/modules/list/actions/toggleTodoCompleted.js
