+++ packages/cerebral-todomvc/src/modules/app/modules/list/chains/removeTodo.js
