+++ packages/cerebral-todomvc/src/modules/app/modules/list/chains/stopEditingTodo.js
