+++ packages/cerebral-todomvc/src/modules/app/modules/list/chains/toggleAllChecked.js
