+++ packages/cerebral-todomvc/src/modules/app/modules/new/actions/setTitle.js
